refactor: optimize json calcluation
