+++ src/format_text.rs
@@ -14,9 +14,9 @@ use super::generation::generate;
 
 pub fn format_text(path: &Path, text: &str, config: &Configuration) -> FormatResult {
   let parse_result = parse(text)?;
-
+  let is_jsonc = is_jsonc_file(path);
   let result = dprint_core::formatting::format(
-    || generate(parse_result, path, text, config),
+    || generate(parse_result, text, config, is_jsonc),
     config_to_print_options(text, config),
   );
   if result == text {
@@ -24,6 +24,56 @@ pub fn format_text(path: &Path, text: &str, config: &Configuration) -> FormatRes
   } else {
     Ok(Some(result))
   }
+}
+
+/// JSONC stands for "JSON with Comments". It is a file specification created by Microsoft and documented here:
+/// https://code.visualstudio.com/docs/languages/json#_json-with-comments
+/// The official parser is written in TypeScript and is located here:
+/// https://github.com/Microsoft/node-jsonc-parser
+/// One of the biggest benefits of JSONC is that it allows trailing commas. Thus, it is desirable to format JSONC with
+/// trailing commas for all the same reasons that code formatters format other languages with trailing commas.
+fn is_jsonc_file(path: &Path) -> bool {
+  return has_jsonc_extension(path) || is_special_json_file(path);
+}
+
+fn has_jsonc_extension(path: &Path) -> bool {
+  if let Some(ext) = path.extension() {
+    return ext.to_str() == Some("jsonc");
+  }
+
+  false
+}
+
+static SPECIAL_JSON_FILES: [&str; 1] = ["tsconfig.json"];
+static SPECIAL_JSON_DIRECTORIES: [&str; 1] = [".vscode"];
+
+/// Some JSONC files use ".json" as a file extension. The best example of this is "tsconfig.json", which is the
+/// configuration file for the TypeScript programming language. When viewing files in VSCode, the language specifier in
+/// the bottom-right corner normally matches what the file extension is. For example, when viewing this file
+/// (generate.rs) in VSCode, the language specifier says "Rust". And when viewing "foo.json" in VSCode, the language
+/// specifier says "JSON". But when viewing "tsconfig.json" in VSCode, the language specifier says "JSON with
+/// Comments". Thus, we must whitelist JSON files with specific paths as being "special" JSON files that should be
+/// treated as JSONC.
+fn is_special_json_file(path: &Path) -> bool {
+  if let Some(file_name) = path.file_name() {
+    if let Some(file_name_str) = file_name.to_str() {
+      if SPECIAL_JSON_FILES.contains(&file_name_str) {
+        return true;
+      }
+    }
+  }
+
+  if let Some(parent_dir) = path.parent() {
+    if let Some(dir_name) = parent_dir.file_name() {
+      if let Some(dir_name_str) = dir_name.to_str() {
+        if SPECIAL_JSON_DIRECTORIES.contains(&dir_name_str) {
+          return true;
+        }
+      }
+    }
+  }
+
+  return false;
 }
 
 #[cfg(feature = "tracing")]
